requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.8
coverage==7.6.12
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0